Add attribute to set DNS name for haproxy admin IP address (bsc#927469)
This is the exact same thing we're doing for the public IP address/DNS
name.

The reason it's needed is for keystone admin API which are being done
through the admin endpoint. As this is a rather minimal use case, we do
not expose the setting in the UI.

https://bugzilla.suse.com/show_bug.cgi?id=927469
(cherry picked from commit 04d82f3dd50e2ed629f1f646cca2d278878fe5cd)