crowbar-pacemaker: Only enable drbd after we've configured resources
In SLE12, drbd fails to start if there are no resources. So we want to
make sure that the resources are defined first. Of course, to configure
resources, we still need the packages to be installed and the global
config, so we now include these drbd recipes (which we were using
indirectly through drbd::default before).

Note that the start itself actually has no real effect as there's a
restart when we configure the resources; however, we still want to
enable the drbd service.